refactor: separate hilt modules
